refactor(event-log): share consequence event-merge across native/WASM + naming fixes (review pass 1)
Item 1: hoist the duplicated Source-1 projection + Source-2 buffer-gate merge
into scp-protocol::trust::consequence::merge_consequence_events (the 3 buffer
constants, the EventType->bucket projection, the four buffer guards, and the
CONVERGENCE INVARIANT doc now live there once). Native
event_log_entries_for_consequences and WASM merged_consequence_events delegate
to it. Behavior-preserving; the only cross-impl difference removed is native's
gappy Source-1 sequence numbering, unified to WASM's dense numbering (sequence
is evidence-only metadata, not consulted by matches_trigger). EL01 + consequence
+ cross-impl leaf-byte parity tests pass unchanged.

Item 2: rename verify_merkle_chain -> recompute_event_log_root (it returns an
RFC-6962 tree root, not a bool, over a tree, not a chain).

Item 3: remove the redundant unsigned ContextExport.merkle_root envelope mirror
and its step-6 consistency check. The signed snapshot.event_log_merkle_root is
the sole authoritative binding (step 5); the envelope copy was always set equal
to it and only re-compared against it, adding ~zero security an attacker could
not trivially satisfy.

Item 4: pruning.rs EXPECTED-table comment 76 -> 75 (matches the 75-entry array).

Item 5: add #[ignore]'d two_real_members_converge_pending_cross_member_replication
pinning that current convergence tests hand-feed identical committer-assigned
timestamps; a real two-member commit-exchange test awaits cross-member leaf
replication (ADR-051 forward step).
